+++ examples/builder/internal-seekable.rb
@@ -18,7 +18,7 @@ module JSONParser
           block.call(json, **options)
         end
       end
-      .terminater do
+      .terminator do
         true
       end
       .build
@@ -38,7 +38,7 @@ module XMLParser
           block.call(REXML::Document.new(file), **options)
         end
       end
-      .terminater do
+      .terminator do
         true
       end
       .build
@@ -79,12 +79,12 @@ module XMLScraper
   end
 end
 
-tree = CallableTree::Node::Root.new.append(
-  JSONParser.build.new.append(
+tree = CallableTree::Node::Root.new.seekable.append(
+  JSONParser.build.new.seekable.append(
     JSONScraper.build(:animals).new,
     JSONScraper.build(:fruits).new
   ),
-  XMLParser.build.new.append(
+  XMLParser.build.new.seekable.append(
     XMLScraper.build(:animals).new,
     XMLScraper.build(:fruits).new
   )